+++ CHANGES.md
@@ -1,5 +1,10 @@
 ### Release Notes
 
+#### **v0.15.9:**
+- improved kernel and notification widget alignment
+- added helvetica, helveticaneue fonts
+- fixed bug that prevented install if fontname not recognized
+
 #### **v0.15.8:**
 - made changes to solarizedd style